sidebar component created
