+++ components/base_layout.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 
 import Head from 'next/head';
-import Link from 'next/link';
 import Script from 'next/script';
 import { useRouter } from 'next/router';
 
@@ -9,84 +8,11 @@ import '@fortawesome/fontawesome-svg-core/styles.css';
 
 import { Footer } from '@components/footer';
 import { Header } from '@components/header';
-import styles from 'styles/common.module.scss';
 
-import {
-    GOOGLE_ANALYTICS_TRACKING_ID,
-    SITE_TITLE,
-    SLUGS_BY_NAME,
-} from '@constants/seo.js';
+import { GOOGLE_ANALYTICS_TRACKING_ID, SITE_TITLE } from '@constants/seo.js';
+import { SidebarList } from '@components/sidebar';
 
 import THEOLOGY101_DATA from '@data/theology101.json';
-
-function getFirstVideoIdForTags(tag) {
-    const videoForTag = Object.values(THEOLOGY101_DATA.lookups.video_id).find(
-        (video) => video.tags && video.tags.includes(tag)
-    );
-    return videoForTag ? videoForTag.videoId : null;
-}
-
-function getFirstVideoIdForTopics(topic) {
-    const videoForTopic = Object.values(THEOLOGY101_DATA.lookups.video_id).find(
-        (video) => video.topic === topic
-    );
-    return videoForTopic ? videoForTopic.videoId : null;
-}
-
-function buildThumbnail(title, setThumbnailUrls) {
-    const fetchThumbnails = async () => {
-        const thumbnailForSidebar = THEOLOGY101_DATA.lookups.video_id;
-        const data =
-            title == 'Topics'
-                ? THEOLOGY101_DATA.lookups.topic
-                : THEOLOGY101_DATA.lookups.tag;
-
-        Object.keys(data).forEach(async (key) => {
-            const videoId =
-                title === 'Topics'
-                    ? getFirstVideoIdForTopics(key)
-                    : getFirstVideoIdForTags(key);
-
-            if (videoId) {
-                setThumbnailUrls((prevUrls) => ({
-                    ...prevUrls,
-                    [key]: thumbnailForSidebar[videoId].thumbnail,
-                }));
-            }
-        });
-    };
-    fetchThumbnails();
-}
-
-function SidebarList({ title, entries, url }) {
-    const [thumbnailUrls, setThumbnailUrls] = React.useState({});
-    React.useEffect(() => {
-        buildThumbnail(title, setThumbnailUrls);
-    }, [title]);
-
-    return (
-        <div>
-            <h5 className="mt-4">{title}</h5>
-            <ul className="list-unstyled">
-                {entries.map(({ name, count }) => (
-                    <li key={name} className={styles.list}>
-                        <Link href={`/${url}/${SLUGS_BY_NAME[url][name]}`}>
-                            <a className="link-secondary text-decoration-none d-flex align-items-center flex-shrink-0">
-                                <img
-                                    className="w-25"
-                                    src={thumbnailUrls[name]}
-                                />
-                                <span className="ms-2 d-flex flex-direction-column">
-                                    <p>{name}</p> <p>({count})</p>
-                                </span>
-                            </a>
-                        </Link>
-                    </li>
-                ))}
-            </ul>
-        </div>
-    );
-}
 
 function buildList(key, lookup) {
     return THEOLOGY101_DATA[key]
